+++ src/main/java/com/yuuko/core/commands/interaction/commands/KissCommand.java
@@ -25,7 +25,7 @@ public class KissCommand extends Command {
 
     @Override
     public void onCommand(MessageReceivedEvent e, String[] command) {
-        Member target = MessageUtilities.getFirstMentionedMember(e);
+        Member target = MessageUtilities.getMentionedMember(e, null, true);
         if(target != null) {
             EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** kisses **" + target.getEffectiveName() + "**.").setImage(interactionImage[new Random().nextInt(interactionImage.length -1)]);
             MessageHandler.sendMessage(e, embed.build());